# docs/openapi.yaml
openapi: 3.1.0
info:
  contact:
    name: DPO оператора
  description: "REST API центра управления согласиями субъектов персональных данны\
    х. Идентификаторы — на английском, пользовательские тексты — на русском (NFR-8)."
  license:
    name: Proprietary
  title: In consensu — центр управления согласиями
  version: v1
servers:
- description: Текущий хост
  url: /
paths:
...
  /api/v1/consent-types/{code}/deactivate:
    post:
      operationId: deactivate_3
      parameters:
      - in: path
        name: code
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ConsentTypeResponse"
          description: OK
      tags:
      - consent-type-controller
...
  /api/v1/notification-rules/{id}/deactivate:
    post:
      operationId: deactivate_2
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/NotificationRuleResponse"
          description: OK
      tags:
      - notification-rule-controller
...
  /api/v1/third-parties/{id}/deactivate:
    post:
      operationId: deactivate_1
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ThirdPartyResponse"
          description: OK
      tags:
      - third-party-controller
...
  /api/v1/webhooks:
    get:
      operationId: list
      responses:
        "200":
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WebhookSubscriptionResponse"
          description: OK
      tags:
      - webhook-controller
    post:
      description: "Секрет подписи возвращается один раз, в ответе на создание"
      operationId: create
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubscriptionRequest"
        required: true
      responses:
        "201":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CreatedSubscriptionResponse"
          description: Created
      summary: Создать подписку
      tags:
      - webhook-controller
  /api/v1/webhooks/{id}:
    get:
      operationId: get
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/WebhookSubscriptionResponse"
          description: OK
      tags:
      - webhook-controller
    put:
      operationId: update
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubscriptionRequest"
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/WebhookSubscriptionResponse"
          description: OK
      tags:
      - webhook-controller
  /api/v1/webhooks/{id}/deactivate:
    post:
      operationId: deactivate
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/WebhookSubscriptionResponse"
          description: OK
      tags:
      - webhook-controller
...
  /api/v1/webhooks/{id}/rotate-secret:
    post:
      description: Новый секрет возвращается один раз
      operationId: rotateSecret
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CreatedSubscriptionResponse"
          description: OK
      summary: Заменить секрет подписи
      tags:
      - webhook-controller
  /api/v1/webhooks/{id}/test:
    post:
      description: Отправляет событие test.ping и возвращает результат попытки
      operationId: test
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/WebhookDeliveryResponse"
          description: OK
      summary: Тестовая отправка (FR-9.5)
      tags:
      - webhook-controller
components:
  schemas:
...
    ConsentTypeResponse:
      type: object
      properties:
        active:
          type: boolean
        businessSignificant:
          type: boolean
        category:
          type: string
        categoryRu:
          type: string
        channels:
          type: array
          items:
            type: string
        code:
          type: string
        defaultValidity:
          type: string
        dependsOnCode:
          type: string
        description:
          type: string
        id:
          type: string
          format: uuid
        nameRu:
          type: string
        requiresThirdParty:
          type: boolean
        sortOrder:
          type: integer
          format: int32
...
    CreatedSubscriptionResponse:
      type: object
      properties:
        secret:
          type: string
        subscription:
          $ref: "#/components/schemas/WebhookSubscriptionResponse"
...
    NotificationRuleResponse:
      type: object
      properties:
        active:
          type: boolean
        channels:
          type: array
          items:
            type: string
        consentTypeId:
          type: string
          format: uuid
        daysBefore:
          type: array
          items:
            type: integer
            format: int32
        id:
          type: string
          format: uuid
        name:
          type: string
        recipientEmails:
          type: array
          items:
            type: string
        recipientRoles:
          type: array
          items:
            type: string
        thirdPartyId:
          type: string
          format: uuid
        triggerType:
          type: string
        triggerTypeRu:
          type: string
...
    SubscriptionRequest:
      type: object
      properties:
        active:
          type: boolean
        eventTypes:
          type: array
          items:
            type: string
          uniqueItems: true
        headers:
          type: object
          additionalProperties:
            type: string
        name:
          type: string
          maxLength: 255
          minLength: 0
        url:
          type: string
          maxLength: 1024
          minLength: 0
      required:
      - name
      - url
...
    ThirdPartyResponse:
      type: object
      properties:
        active:
          type: boolean
        address:
          type: string
        allowedPdnCategories:
          type: array
          items:
            type: string
        contactEmail:
          type: string
        contractDate:
          type: string
          format: date
        contractDaysLeft:
          type: integer
          format: int64
        contractExpired:
          type: boolean
        contractNumber:
          type: string
        contractValidUntil:
          type: string
          format: date
        id:
          type: string
          format: uuid
        inn:
          type: string
        name:
          type: string
        ogrn:
          type: string
        role:
          type: string
        roleRu:
          type: string
        shortName:
          type: string
...
    WebhookDeliveryResponse:
      type: object
      properties:
        attempt:
          type: integer
          format: int32
        deliveredAt:
          type: string
          format: date-time
        deliveryId:
          type: string
          format: uuid
        error:
          type: string
        id:
          type: string
          format: uuid
        responseCode:
          type: integer
          format: int32
        subscriptionId:
          type: string
          format: uuid
        successful:
          type: boolean
    WebhookSubscriptionResponse:
      type: object
      properties:
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
        eventTypes:
          type: array
          items:
            type: string
        id:
          type: string
          format: uuid
        name:
          type: string
        updatedAt:
          type: string
          format: date-time
        url:
          type: string